address_validation/component.py
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
#
# Permission is hereby granted, free of charge, to any person obtaining a copy of
# this software and associated documentation files (the "Software"), to deal in
# the Software without restriction, including without limitation the rights to
# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
# the Software, and to permit persons to whom the Software is furnished to do so.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

from typing import Any

import aws_cdk as cdk
import aws_cdk.aws_dynamodb as dynamodb
from constructs import Construct

from address_validation.datapipeline.infrastructure import DataPipeline as Avdp


class AddressValidation(cdk.Stack):
    def __init__(
        self,
        scope: Construct,
        id_:str,
        **kwargs: Any,
    ):
        super().__init__(scope, id_, **kwargs)

        # TODO Edit below to build the pip
        val_data_pipeline = Avdp(
            self,
            "DataPipeline",
            account = self.account,
            region = self.region
        )

        # TODO Edit below to add outputs
        self.api_endpoint = cdk.CfnOutput(
            self,
            "Account",
            value=self.account
        )