requirements-frozen.txt
attrs==23.1.0
aws-cdk-lib==2.84.0
aws-cdk.asset-awscli-v1==2.2.197
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.165
aws-cdk.aws-glue-alpha==2.84.0a0
awswrangler==3.0.0
bandit==1.7.5
boto3==1.26.133
botocore==1.29.133
cattrs==22.2.0
cdk-nag==2.27.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
constructs==10.2.19
decorator==5.1.1
exceptiongroup==1.1.1
future==0.18.3
geocoder==1.38.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-resources==5.12.0
install==1.3.5
jmespath==1.0.1
jsii==1.84.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.24.3
pandas==1.5.3
pbr==5.11.1
probableparsing==0.0.1
publication==0.0.3
pyarrow==12.0.0
Pygments==2.15.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
requests==2.31.0
rich==13.4.2
s3transfer==0.6.1
six==1.16.0
smartystreets-python-sdk==4.11.16
smmap==5.0.0
stevedore==5.1.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
usaddress==0.5.10
usaddress-scourgify==0.5.1
yaml-config==0.1.5
zipp==3.15.0
